VBM on Android - Top Stories Screen : Story Network available

+++ HackerNewsReader/src/main/java/com/quantityandconversion/hackernews/network/hackernews/internal/HackerNewsApi.java
@@ -14,5 +14,5 @@ import retrofit2.http.Path;
     Call<Stories> topStories();
 
     @GET("item/{storyId}.json")
-    Call<Story> story(@Path("storyId") final int storyId);
+    Call<Story> story(@Path("storyId") final long storyId);
 }
